+++ MapControl/MapBase.Silverlight.WinRT.cs
@@ -71,7 +71,7 @@ namespace MapControl
         private void SetViewportTransform(Location origin)
         {
             MapOrigin = mapTransform.Transform(origin);
-            ViewportScale = Math.Pow(2d, ZoomLevel) * (double)TileSource.TileSize / 360d;
+            ViewportScale = Math.Pow(2d, ZoomLevel) * TileSource.TileSize / 360d;
 
             var transform = new Matrix(1d, 0d, 0d, 1d, -MapOrigin.X, -MapOrigin.Y)
                 .Rotate(-Heading)
