+++ MapControl/TileLayer.WPF.cs
@@ -18,7 +18,7 @@ namespace MapControl
 
         private Rect GetTileIndexBounds(int zoomLevel)
         {
-            var scale = (double)(1 << zoomLevel) / 360d;
+            var scale = (1 << zoomLevel) / 360d;
             var transform = parentMap.ViewportTransform.Matrix;
 
             transform.Invert(); // view to map coordinates
