+++ MapControl/TileSource.cs
@@ -160,11 +160,11 @@ namespace MapControl
 
         private Uri GetBoundingBoxUri(int x, int y, int zoomLevel)
         {
-            var tileSize = 360d / (double)(1 << zoomLevel); // tile width in degrees
-            var west = MetersPerDegree * ((double)x * tileSize - 180d);
-            var east = MetersPerDegree * ((double)(x + 1) * tileSize - 180d);
-            var south = MetersPerDegree * (180d - (double)(y + 1) * tileSize);
-            var north = MetersPerDegree * (180d - (double)y * tileSize);
+            var tileSize = 360d / (1 << zoomLevel); // tile width in degrees
+            var west = MetersPerDegree * (x * tileSize - 180d);
+            var east = MetersPerDegree * ((x + 1) * tileSize - 180d);
+            var south = MetersPerDegree * (180d - (y + 1) * tileSize);
+            var north = MetersPerDegree * (180d - y * tileSize);
 
             return new Uri(uriFormat
                 .Replace("{W}", west.ToString(CultureInfo.InvariantCulture))
@@ -177,11 +177,11 @@ namespace MapControl
 
         private Uri GetLatLonBoundingBoxUri(int x, int y, int zoomLevel)
         {
-            var tileSize = 360d / (double)(1 << zoomLevel); // tile width in degrees
-            var west = (double)x * tileSize - 180d;
-            var east = (double)(x + 1) * tileSize - 180d;
-            var south = MercatorTransform.YToLatitude(180d - (double)(y + 1) * tileSize);
-            var north = MercatorTransform.YToLatitude(180d - (double)y * tileSize);
+            var tileSize = 360d / (1 << zoomLevel); // tile width in degrees
+            var west = x * tileSize - 180d;
+            var east = (x + 1) * tileSize - 180d;
+            var south = MercatorTransform.YToLatitude(180d - (y + 1) * tileSize);
+            var north = MercatorTransform.YToLatitude(180d - y * tileSize);
 
             return new Uri(uriFormat
                 .Replace("{w}", west.ToString(CultureInfo.InvariantCulture))
